+++ client/src/components/App.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import NavBar from './NavBar';
+import Greeter from './Greeter';
 import About from './About';
 import Projects from './Projects';
 import Expertise from './Expertise';
 import Footer from './Footer';
 
 const App = () => (
-  <div className="section">
-    <NavBar />
+  <div>
+    <Greeter />
     <About />
     <Projects />
     <Expertise />
